refactor: modularize generations.export
