--- src/server/api/routers/generations.ts
@@ -1,621 +1,0 @@
-import { z } from "zod";
-
-import {
-  createTRPCRouter,
-  protectedProjectProcedure,
-} from "@/src/server/api/trpc";
-
-import { Prisma, type ObservationView } from "@prisma/client";
-import { jsonSchema, paginationZod } from "@/src/utils/zod";
-import { singleFilter } from "@/src/server/api/interfaces/filters";
-import {
-  datetimeFilterToPrismaSql,
-  filterToPrismaSql,
-} from "@/src/features/filters/server/filterToPrisma";
-import {
-  type ObservationOptions,
-  observationsTableCols,
-} from "@/src/server/api/definitions/observationsTable";
-import { usdFormatter } from "@/src/utils/numbers";
-import { env } from "@/src/env.mjs";
-import {
-  S3Client,
-  PutObjectCommand,
-  GetObjectCommand,
-} from "@aws-sdk/client-s3";
-import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
-import {
-  exportFileFormats,
-  exportOptions,
-} from "@/src/server/api/interfaces/exportTypes";
-import { orderBy } from "@/src/server/api/interfaces/orderBy";
-import { orderByToPrismaSql } from "@/src/features/orderBy/server/orderByToPrisma";
-
-const GenerationTableOptions = z.object({
-  projectId: z.string(), // Required for protectedProjectProcedure
-  filter: z.array(singleFilter),
-  searchQuery: z.string().nullable(),
-  orderBy: orderBy,
-});
-
-const ListInputs = GenerationTableOptions.extend({
-  ...paginationZod,
-});
-
-// extend generationfilteroptions with export options
-const ExportInputs = GenerationTableOptions.extend({
-  fileFormat: z.enum(exportFileFormats),
-});
-
-export const generationsRouter = createTRPCRouter({
-  all: protectedProjectProcedure
-    .input(ListInputs)
-    .query(async ({ input, ctx }) => {
-      // ATTENTION: When making changes to this query, make sure to also update the export query
-      const searchCondition = input.searchQuery
-        ? Prisma.sql`AND (
-        o."id" ILIKE ${`%${input.searchQuery}%`} OR
-        o."name" ILIKE ${`%${input.searchQuery}%`} OR
-        o."model" ILIKE ${`%${input.searchQuery}%`} OR
-        t."name" ILIKE ${`%${input.searchQuery}%`}
-      )`
-        : Prisma.empty;
-
-      const filterCondition = filterToPrismaSql(
-        input.filter,
-        observationsTableCols,
-      );
-
-      const orderByCondition = orderByToPrismaSql(
-        input.orderBy,
-        observationsTableCols,
-      );
-
-      // to improve query performance, add timeseries filter to observation queries as well
-      const startTimeFilter = input.filter.find(
-        (f) => f.column === "start_time" && f.type === "datetime",
-      );
-      const datetimeFilter =
-        startTimeFilter && startTimeFilter.type === "datetime"
-          ? datetimeFilterToPrismaSql(
-              "start_time",
-              startTimeFilter.operator,
-              startTimeFilter.value,
-            )
-          : Prisma.empty;
-
-      const generations = await ctx.prisma.$queryRaw<
-        Array<
-          ObservationView & {
-            traceId: string;
-            traceName: string;
-            latency: number | null;
-          }
-        >
-      >(
-        Prisma.sql`
-          WITH observations_with_latency AS (
-            SELECT
-              o.*,
-              CASE WHEN o.end_time IS NULL THEN NULL ELSE (EXTRACT(EPOCH FROM o."end_time") - EXTRACT(EPOCH FROM o."start_time"))::double precision END AS "latency"
-            FROM observations_view o
-            WHERE o.type = 'GENERATION'
-            AND o.project_id = ${input.projectId}
-            ${datetimeFilter}
-          ),
-          -- used for filtering
-          scores_avg AS (
-            SELECT
-              trace_id,
-              observation_id,
-              jsonb_object_agg(name::text, avg_value::double precision) AS scores_avg
-            FROM (
-              SELECT
-                trace_id,
-                observation_id,
-                name,
-                avg(value) avg_value
-              FROM
-                scores
-              GROUP BY
-                1,
-                2,
-                3
-              ORDER BY
-                1) tmp
-            GROUP BY
-              1, 2
-          )
-          SELECT
-            o.id,
-            o.name,
-            o.model,
-            o.start_time as "startTime",
-            o.end_time as "endTime",
-            o.latency,
-            o.input,
-            o.output,
-            o.metadata,
-            o.trace_id as "traceId",
-            t.name as "traceName",
-            o.completion_start_time as "completionStartTime",
-            o.prompt_tokens as "promptTokens",
-            o.completion_tokens as "completionTokens",
-            o.total_tokens as "totalTokens",
-            o.level,
-            o.status_message as "statusMessage",
-            o.version,
-            o.model_id as "modelId",
-            o.input_price as "inputPrice",
-            o.output_price as "outputPrice",
-            o.total_price as "totalPrice",
-            o.calculated_input_cost as "calculatedInputCost",
-            o.calculated_output_cost as "calculatedOutputCost",
-            o.calculated_total_cost as "calculatedTotalCost"
-          FROM observations_with_latency o
-          JOIN traces t ON t.id = o.trace_id
-          LEFT JOIN scores_avg AS s_avg ON s_avg.trace_id = t.id and s_avg.observation_id = o.id
-          WHERE
-            t.project_id = ${input.projectId}
-            ${searchCondition}
-            ${filterCondition}
-            ${orderByCondition}
-          LIMIT ${input.limit}
-          OFFSET ${input.page * input.limit}
-        `,
-      );
-
-      const totalGenerations = await ctx.prisma.$queryRaw<
-        Array<{ count: bigint }>
-      >(
-        Prisma.sql`
-          WITH observations_with_latency AS (
-            SELECT
-              o.*,
-              CASE WHEN o.end_time IS NULL THEN NULL ELSE (EXTRACT(EPOCH FROM o."end_time") - EXTRACT(EPOCH FROM o."start_time"))::double precision END AS "latency"
-            FROM observations_view o
-            WHERE o.type = 'GENERATION'
-            AND o.project_id = ${input.projectId}
-            ${datetimeFilter}
-          ),
-          -- used for filtering
-          scores_avg AS (
-            SELECT
-              trace_id,
-              observation_id,
-              jsonb_object_agg(name::text, avg_value::double precision) AS scores_avg
-            FROM (
-              SELECT
-                trace_id,
-                observation_id,
-                name,
-                avg(value) avg_value
-              FROM
-                scores
-              GROUP BY
-                1,
-                2,
-                3
-              ORDER BY
-                1) tmp
-            GROUP BY
-              1, 2
-          )
-          SELECT
-            count(*)
-          FROM observations_with_latency o
-          JOIN traces t ON t.id = o.trace_id
-          LEFT JOIN scores_avg AS s_avg ON s_avg.trace_id = t.id and s_avg.observation_id = o.id
-          WHERE
-            t.project_id = ${input.projectId}
-            ${searchCondition}
-            ${filterCondition}
-        `,
-      );
-
-      const scores = await ctx.prisma.score.findMany({
-        where: {
-          trace: {
-            projectId: input.projectId,
-          },
-          observationId: {
-            in: generations.map((gen) => gen.id),
-          },
-        },
-      });
-      const count = totalGenerations[0]?.count;
-      return {
-        totalCount: count ? Number(count) : undefined,
-        generations: generations.map((generation) => {
-          const filteredScores = scores.filter(
-            (s) => s.observationId === generation.id,
-          );
-          return {
-            ...generation,
-            scores: filteredScores,
-          };
-        }),
-      };
-    }),
-
-  export: protectedProjectProcedure
-    .input(ExportInputs)
-    .query(async ({ input, ctx }) => {
-      // ATTENTION: When making changes to this query, make sure to also update the all query
-      const searchCondition = input.searchQuery
-        ? Prisma.sql`AND (
-        o."id" ILIKE ${`%${input.searchQuery}%`} OR
-        o."name" ILIKE ${`%${input.searchQuery}%`} OR
-        o."model" ILIKE ${`%${input.searchQuery}%`} OR
-        t."name" ILIKE ${`%${input.searchQuery}%`}
-      )`
-        : Prisma.empty;
-
-      const filterCondition = filterToPrismaSql(
-        input.filter,
-        observationsTableCols,
-      );
-
-      const orderByCondition = orderByToPrismaSql(
-        input.orderBy,
-        observationsTableCols,
-      );
-
-      // to improve query performance, add timeseries filter to observation queries as well
-      const startTimeFilter = input.filter.find(
-        (f) => f.column === "start_time" && f.type === "datetime",
-      );
-      const datetimeFilter =
-        startTimeFilter && startTimeFilter.type === "datetime"
-          ? datetimeFilterToPrismaSql(
-              "start_time",
-              startTimeFilter.operator,
-              startTimeFilter.value,
-            )
-          : Prisma.empty;
-
-      const allGenerations = await ctx.prisma.$queryRaw<
-        Array<
-          ObservationView & {
-            traceId: string;
-            traceName: string;
-            latency: number | null;
-          }
-        >
-      >(
-        Prisma.sql`
-          WITH observations_with_latency AS (
-            SELECT
-              o.*,
-              CASE WHEN o.end_time IS NULL THEN NULL ELSE (EXTRACT(EPOCH FROM o."end_time") - EXTRACT(EPOCH FROM o."start_time"))::double precision END AS "latency"
-            FROM observations_view o
-            WHERE o.type = 'GENERATION'
-            AND o.project_id = ${input.projectId}
-            ${datetimeFilter}
-          ),
-          -- used for filtering
-          scores_avg AS (
-            SELECT
-              trace_id,
-              observation_id,
-              jsonb_object_agg(name::text, avg_value::double precision) AS scores_avg
-            FROM (
-              SELECT
-                trace_id,
-                observation_id,
-                name,
-                avg(value) avg_value
-              FROM
-                scores
-              GROUP BY
-                1,
-                2,
-                3
-              ORDER BY
-                1) tmp
-            GROUP BY
-              1, 2
-          )
-          SELECT
-            o.id,
-            o.name,
-            o.model,
-            o.start_time as "startTime",
-            o.end_time as "endTime",
-            o.latency,
-            o.input,
-            o.output,
-            o.metadata,
-            o.trace_id as "traceId",
-            t.name as "traceName",
-            o.completion_start_time as "completionStartTime",
-            o.prompt_tokens as "promptTokens",
-            o.completion_tokens as "completionTokens",
-            o.total_tokens as "totalTokens",
-            o.level,
-            o.status_message as "statusMessage",
-            o.version,
-            o.model_id as "modelId",
-            o.input_price as "inputPrice",
-            o.output_price as "outputPrice",
-            o.total_price as "totalPrice",
-            o.calculated_input_cost as "calculatedInputCost",
-            o.calculated_output_cost as "calculatedOutputCost",
-            o.calculated_total_cost as "calculatedTotalCost"
-          FROM observations_with_latency o
-          JOIN traces t ON t.id = o.trace_id
-          LEFT JOIN scores_avg AS s_avg ON s_avg.trace_id = t.id and s_avg.observation_id = o.id
-          WHERE
-            t.project_id = ${input.projectId}
-            ${searchCondition}
-            ${filterCondition}
-            ${orderByCondition}
-        `,
-      );
-
-      let output: string = "";
-
-      // Create file string
-      switch (input.fileFormat) {
-        case "CSV":
-          output = mutateGenerationsInPlaceToCSV(allGenerations);
-          break;
-        case "JSON":
-          // This operation is not yet memory optimized to mutate the generations in place
-          output = JSON.stringify(allGenerations);
-          break;
-        case "OPENAI-JSONL":
-          output = mutateGenerationsInPlaceToJSONL(allGenerations);
-
-          break;
-        default:
-          throw new Error("Invalid export file format");
-      }
-
-      const fileDate = new Date().toISOString();
-      const fileExtension = exportOptions[input.fileFormat].extension;
-      const fileName = `lf-export-${input.projectId}-${fileDate}.${fileExtension}`;
-
-      // Upload to S3 if credentials are provided
-      if (
-        env.S3_BUCKET_NAME &&
-        env.S3_ACCESS_KEY_ID &&
-        env.S3_SECRET_ACCESS_KEY &&
-        env.S3_ENDPOINT &&
-        env.S3_REGION
-      ) {
-        const client = new S3Client({
-          credentials: {
-            accessKeyId: env.S3_ACCESS_KEY_ID,
-            secretAccessKey: env.S3_SECRET_ACCESS_KEY,
-          },
-          endpoint: env.S3_ENDPOINT,
-          region: env.S3_REGION,
-        });
-
-        await client.send(
-          new PutObjectCommand({
-            Bucket: env.S3_BUCKET_NAME,
-            Key: fileName,
-            Body: output,
-            ContentType: exportOptions[input.fileFormat].fileType,
-          }),
-        );
-
-        const signedUrl = await getSignedUrl(
-          client,
-          new GetObjectCommand({
-            Bucket: env.S3_BUCKET_NAME,
-            Key: fileName,
-            ResponseContentDisposition: `attachment; filename="${fileName}"`,
-          }),
-          {
-            expiresIn: 60 * 60, // in 1 hour, signed url will expire
-          },
-        );
-
-        return {
-          type: "s3",
-          url: signedUrl,
-          fileName,
-        } as const;
-      }
-
-      // Fallback to returning the file as a data in HTTP response
-      return {
-        type: "data",
-        data: output,
-        fileName,
-      } as const;
-    }),
-
-  filterOptions: protectedProjectProcedure
-    .input(z.object({ projectId: z.string() }))
-    .query(async ({ input, ctx }) => {
-      const queryFilter = {
-        projectId: input.projectId,
-        type: "GENERATION",
-      } as const;
-
-      const scores = await ctx.prisma.score.groupBy({
-        where: {
-          observation: {
-            projectId: input.projectId,
-          },
-        },
-        by: ["name"],
-      });
-
-      const model = await ctx.prisma.observation.groupBy({
-        by: ["model"],
-        where: queryFilter,
-        _count: { _all: true },
-      });
-      const name = await ctx.prisma.observation.groupBy({
-        by: ["name"],
-        where: queryFilter,
-        _count: { _all: true },
-      });
-      const traceName = await ctx.prisma.$queryRaw<
-        Array<{
-          traceName: string | null;
-          count: number;
-        }>
-      >(Prisma.sql`
-        SELECT
-          t.name "traceName",
-          count(*)::int AS count
-        FROM traces t
-        JOIN observations o ON o.trace_id = t.id
-        WHERE o.type = 'GENERATION'
-          AND o.project_id = ${input.projectId}
-          AND t.project_id = ${input.projectId}
-        GROUP BY 1
-      `);
-
-      // typecheck filter options, needs to include all columns with options
-      const res: ObservationOptions = {
-        model: model
-          .filter((i) => i.model !== null)
-          .map((i) => ({
-            value: i.model as string,
-            count: i._count._all,
-          })),
-        name: name
-          .filter((i) => i.name !== null)
-          .map((i) => ({
-            value: i.name as string,
-            count: i._count._all,
-          })),
-        traceName: traceName
-          .filter((i) => i.traceName !== null)
-          .map((i) => ({
-            value: i.traceName as string,
-            count: i.count,
-          })),
-        scores_avg: scores.map((score) => score.name),
-      };
-      return res;
-    }),
-});
-
-/**
- * WARNING: Mutates the allGenerations array in place.
- * Converts an array of ObservationView objects to a CSV string.
- * The array is mutated in place to avoid memory issues with large datasets.
- *
- * @param allGenerations - An array of ObservationView objects.
- * @returns The CSV string representation of the observations.
- * @sideEffects Mutates the allGenerations array in place to an `undefined[]`.
- * @throws Error if the input is invalid.
- */
-function mutateGenerationsInPlaceToCSV(
-  allGenerations: (ObservationView | undefined)[] | string,
-): string {
-  if (typeof allGenerations === "string") {
-    throw Error("Invalid input");
-  }
-
-  const csvHeader = [
-    "traceId",
-    "name",
-    "model",
-    "startTime",
-    "endTime",
-    "cost",
-    "prompt",
-    "completion",
-    "metadata",
-  ];
-
-  let output = csvHeader.join(",");
-
-  allGenerations.forEach((generation, index, allGenerations) => {
-    if (!generation) {
-      return;
-    }
-
-    const csvRow = [
-      generation.traceId,
-      generation.name ?? "",
-      generation.model ?? "",
-      generation.startTime.toISOString(),
-      generation.endTime?.toISOString() ?? "",
-      generation.calculatedTotalCost
-        ? usdFormatter(generation.calculatedTotalCost.toNumber(), 2, 8)
-        : "",
-      JSON.stringify(generation.input),
-      JSON.stringify(generation.output),
-      JSON.stringify(generation.metadata),
-    ].map((field) => {
-      const str = typeof field === "string" ? field : String(field);
-      return `"${str.replace(/"/g, '""')}"`;
-    });
-
-    output += "\n" + csvRow.join(",");
-    allGenerations[index] = undefined;
-  });
-
-  return output;
-}
-
-/**
- * WARNING: Mutates the allGenerations array in place.
- * Converts an array of ObservationView objects to a JSONL string.
- * The array is mutated in place to avoid memory issues with large datasets.
- *
- * @param allGenerations - An array of ObservationView objects.
- * @returns The JSONL string representation of the observations.
- * @sideEffects Mutates the allGenerations array in place to an `undefined[]`.
- * @throws Error if the input is invalid.
- */
-function mutateGenerationsInPlaceToJSONL(
-  allGenerations: (ObservationView | undefined)[] | string,
-): string {
-  if (typeof allGenerations === "string") {
-    throw Error("Invalid input");
-  }
-
-  const inputSchemaOpenAI = z.array(
-    z.object({
-      role: z.enum(["system", "user", "assistant"]),
-      content: z.string(),
-    }),
-  );
-
-  const outputSchema = z
-    .object({
-      completion: jsonSchema,
-    })
-    .or(jsonSchema);
-
-  let output: string = "";
-
-  allGenerations.forEach((generation, index, allGenerations) => {
-    if (!generation) {
-      return;
-    }
-
-    const parsedInput = inputSchemaOpenAI.safeParse(generation.input);
-    const parsedOutput = outputSchema.safeParse(generation.output);
-
-    if (parsedInput.success && parsedOutput.success) {
-      output +=
-        JSON.stringify([
-          ...parsedInput.data,
-          {
-            role: "assistant",
-            content:
-              typeof parsedOutput.data === "object" &&
-              "completion" in parsedOutput.data
-                ? JSON.stringify(parsedOutput.data.completion)
-                : JSON.stringify(parsedOutput.data),
-          },
-        ]) + "\n";
-    }
-
-    allGenerations[index] = undefined;
-  });
-
-  return output;
-}
